Changes to make the code python3 compatible and refactoring to remove circular dependencies
