+++ ample/ensembler/subcluster.py
@@ -13,8 +13,7 @@ import re
 import os
 import shutil
 
-from ample.util import ample_util
-from ample.util import pdb_edit
+from ample.util import ample_util, pdb_edit
 
 logger = logging.getLogger()
