+++ ample/util/exit_util.py
@@ -1,8 +1,8 @@
-'''
+"""
 Created on Mar 18, 2015
 
 @author: jmht
-'''
+"""
 
 import logging
 import os
