+++ ample/util/mrbump_util.py
@@ -17,12 +17,7 @@ from ample.util import ample_util
 from ample.util import mrbump_cmd
 from ample.util import printTable
 
-# FS [15/11/2018] -> New MRBUMP structure allows direct imports, wait for release then replace
-mrbumpd = os.path.join(os.environ['CCP4'], "share", "mrbump", "include", "parsers")
-sys.path.insert(0, mrbumpd)
-import parse_arpwarp
-import parse_buccaneer
-import parse_phaser
+from mrbump.parsers import parse_arpwarp, parse_buccaneer, parse_phaser
 
 TOP_KEEP = 3  # How many of the top shelxe/phaser results to keep for the gui
 MRBUMP_RUNTIME = 172800  # allow 48 hours for each mrbump job
