+++ ample/util/phenixer.py
@@ -1,8 +1,8 @@
-'''
+"""
 Created on 5 Mar 2014
 
 @author: jmht
-'''
+"""
 
 import os
 import shutil
