+++ ample/util/scwrl_util.py
@@ -1,6 +1,6 @@
-'''
+"""
 @author: jmht
-'''
+"""
 
 import glob
 import os
